=comments managed , ACL added

+++ resources/views/admin/brand/index.blade.php
@@ -33,19 +33,23 @@
             <td class="">{{ $brand->brand_name }}</td>
             <td class="">{{ $brand->brand_slug }}</td>
             <td class="center">
-               <img src="{{ $brand->src }}" alt="brand image">
+               <img src="{{ $brand->src }}" alt="brand image" class="img-responsive img-thumbnail" >
             </td>
             <td class="">{{ $brand->brand_description }}</td>
             <td class="center">
                <div class="hidden-sm hidden-xs btn-group">
                   <form>
-                     <button class="btn btn-xs btn-danger delete_me" data-id="{{ $brand->brand_id }}">
-                        <i class="ace-icon fa fa-trash-o bigger-120"></i>
-                     </button>
-                     <a class="btn btn-warning btn-xs edit_me" title="Edit"
-                        href="{{ route('brand.edit',$brand->brand_id) }}" data-id="{{ $brand->brand_id }}">
-                        <i class="ace-icon fa fa-pencil bigger-120"></i>
-                     </a>
+                     @can('product-delete')
+                        <button class="btn btn-xs btn-danger delete_me" data-id="{{ $brand->brand_id }}">
+                           <i class="ace-icon fa fa-trash-o bigger-120"></i>
+                        </button>
+                     @endcan
+                     @can('product-edit')
+                        <a class="btn btn-warning btn-xs click_me" title="Edit"
+                           href="{{ route('brand.edit',$brand->brand_id) }}" data-id="{{ $brand->brand_id }}">
+                           <i class="ace-icon fa fa-pencil bigger-120"></i>
+                        </a>
+                     @endcan
                   </form>
                </div>
             </td>
@@ -57,23 +61,16 @@
       @endforelse
       </tbody>
    </table>
-      {{ $admin_brands->links() }}
+   {{ $admin_brands->links() }}
 
 @endsection
 @section('extra_js')
    <script>
        $(document).ready(function () {
+          @can('product-delete')
            //DELETE ROW
-           deleteAjax("/admin/brand/","delete_me","brand");
-           <!-- LOAD THE EDIT PAGE-->
-           jQuery(".edit_me").bind('click', function () {
-               var route = $(this).attr('href');
-               var pjax = new Pjax({
-                   selectors: ["title", "#extra_css", "#content-load", "#extra_js"]
-               });
-               pjax.loadUrl(route);
-
-           });
+           deleteAjax("/admin/brand/", "delete_me", "brand");
+          @endcan
        });
    </script>
 @stop
